+++ src/platform/plugins/shared/discover/public/application/main/components/session_view/main_app.tsx
