+++ src/platform/plugins/shared/discover/public/application/main/discover_main_route.test.tsx
@@ -16,7 +16,7 @@ import { discoverServiceMock } from '../../__mocks__/services';
 import type { MainRouteProps } from './discover_main_route';
 import { DiscoverMainRoute } from './discover_main_route';
 import { MemoryRouter } from 'react-router-dom';
-import { DiscoverMainApp } from './components/session_view/discover_main_app';
+import { DiscoverMainApp } from './components/session_view/main_app';
 import { findTestSubject } from '@elastic/eui/lib/test';
 import type { DiscoverCustomizationService } from '../../customizations/customization_service';
 import { createCustomizationService } from '../../customizations/customization_service';
@@ -35,7 +35,7 @@ jest.mock('../../customizations', () => {
   };
 });
 
-jest.mock('./components/session_view/discover_main_app', () => {
+jest.mock('./components/session_view/main_app', () => {
   return {
     DiscoverMainApp: jest.fn().mockReturnValue(<></>),
   };
